Namespace search and filter clauses
