--- lib/yuriita/filter_clause.rb
@@ -1,23 +1,0 @@
-module Yuriita
-  class FilterClause
-    def initialize(filters:, combination:)
-      @filters = filters
-      @combination = combination
-    end
-
-    def apply(relation)
-      return relation if unselected?
-
-      relations = filters.map { |filter| filter.apply(relation) }
-      combination.new(base_relation: relation, relations: relations).combine
-    end
-
-    private
-
-    attr_reader :filters, :combination
-
-    def unselected?
-      filters.empty?
-    end
-  end
-end
