+++ lib/yuriita/scope.rb
@@ -10,7 +10,7 @@ module Yuriita
     def apply(query:)
       selector = AllOrExplicitSelect.new(options: options, query: query)
 
-      SearchClause.new(
+      Clauses::Search.new(
         filters: selector.filters,
         keywords: query.keywords,
         combination: combination,
