--- lib/yuriita/search_clause.rb
@@ -1,24 +1,0 @@
-module Yuriita
-  class SearchClause
-    def initialize(filters:, keywords:, combination:)
-      @filters = filters
-      @keywords = keywords
-      @combination = combination
-    end
-
-    def apply(relation)
-      return relation if unselected?
-
-      relations = filters.map { |filter| filter.apply(relation, keywords) }
-      combination.new(base_relation: relation, relations: relations).combine
-    end
-
-    private
-
-    attr_reader :filters, :keywords, :combination
-
-    def unselected?
-      filters.empty? || keywords.empty?
-    end
-  end
-end
